Move code out into two files.
